backend/requirements.txt
fastapi
uvicorn
pydantic-ai
pydantic-settings
pymongo==4.6.1
motor==3.3.2
python-dotenv
websockets
pytest
pytest-asyncio==0.23.3
httpx
ruff==0.1.14
black==24.1.1
email-validator==2.1.0
griffe
# Phase 2 dependencies
apscheduler
requests
